+++ src/controllers/index.js
@@ -1,0 +1,4 @@
+import combineControllers from '@home/core/combineControllers'
+import user from './user.ctrl';
+
+export default combineControllers(user);