+++ src/controllers/user.ctrl.js
@@ -1,0 +1,13 @@
+export default class UserController {
+
+	namespace = 'user';
+
+	initial = {
+		info: {},
+		permission: {},
+	};
+
+	saveUser({ payload }, state) {
+		return payload;
+	}
+}