+++ src/core/baseClass/Service.js
@@ -1,0 +1,5 @@
+export default class Service {
+	// 类型是service
+	$type = 'service';
+	
+}