+++ src/core/launcher/index.js
@@ -2,10 +2,16 @@ import render from './render'
 import createApp from './createApp'
 import getMountableComponent from './getMountableComponent'
 import createStore from './createStore'
+import container from './container'
+import store from './store'
+import Service from './Service'
 
 export default {
 	render,
 	createApp,
 	getMountableComponent,
 	createStore,
+	container,
+	store,
+	Service,
 }