+++ src/epics/index.js
@@ -1,0 +1,5 @@
+import combineEpics from '@home/core/combineEpics'
+
+import user from './user.epic';
+
+export default combineEpics(user);